requirements.txt
fastapi
pydantic
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib
python-jose